refactor: 구조개선
각 객체간의 역할과 메시지 전달 주체를 분명히 하기 위한 개선

1. LottoMachine
- 구매 금액을 입력한다.
- 당첨번호를 입력한다.

2. Lotto
- 로또를 개수만큼 생성한다.
- 번호 노출 여부에 따라 re-render 한다.

3. LottoAnalytics
- 당첨 번호, 구매한 로또 번호, 투자금(구매한 금액)으로 당첨 통계 및 수익률을 계산한다.
- 통계 모달을 on/off 한다.

이외에 개선점 TODO로 기록
