+++ cypress/integration/lotto.spec.js
@@ -262,4 +262,9 @@ describe("로또 계산기", () => {
       expect(earningRates).to.be.eq(100);
     });
   });
+
+  /**
+   * 로또 구매를 완료하면 당첨번호 입력 form이 노출된다.
+   * 당첨번호 입력 완료하면 모달이 뜬다. 등등
+   */
 });
